# openapi/fc_openapi.yaml
openapi: 3.0.1
info:
  title: Eclipse XFSC Federated Catalogue
  description: REST API of the XFSC catalogue
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://fc-server.xfsc.org
    description: some future test environment
components:
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    OntologySchema:
      type: object
      properties:
        ontologies:
          description: 'SchemaIds of ontologies, defining the terms of a schema'
          type: array
          items:
            type: string
            example: schemaId
        shapes:
          description: 'SchemaIds of shapes (SHACL Shapes Constraint Language) used to validate instances against the schema.'
          type: array
          items:
            type: string
            example: schemaId
        vocabularies:
          description: 'SchemaIds of controlled vocabularies used in the schema'
          type: array
          items:
            type: string
            example: schemaId
        jsonSchemas:
          description: 'SchemaIds of JSON Schemas used to validate JSON assets.'
          type: array
          items:
            type: string
            example: schemaId
        xmlSchemas:
          description: 'SchemaIds of XML Schemas (XSD) used to validate XML assets.'
          type: array
          items:
            type: string
            example: schemaId
...
    SchemaResult:
      type: object
      properties:
        id:
          type: string
          description: 'The internal identifier of the schema.'
        warnings:
          type: array
          items:
            type: string
          description: 'Warnings generated during schema import (e.g. filtered protected namespace statements).'
        createdAt:
          type: string
          format: date-time
          nullable: true
          description: 'Timestamp when the schema was first uploaded. Present on creation, null on updates.'
        version:
          type: integer
          nullable: true
          description: 'Current version number after this operation. Present on updates, null on creation.'
        previousVersion:
          type: integer
          nullable: true
          description: 'Version number before this operation. Null for first version or on creation.'
...
    SchemaVersion:
      type: object
      properties:
        version:
          type: integer
          description: '1-based version ordinal.'
        createdAt:
          type: string
          format: date-time
          description: 'Timestamp when this version was created.'
        isCurrent:
          type: boolean
          description: 'Whether this is the current (latest) version.'
    SchemaVersionList:
      type: object
      properties:
        schemaId:
          type: string
          description: 'The schema identifier.'
        versions:
          type: array
          items:
            $ref: '#/components/schemas/SchemaVersion'
...
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: >-
        Forbidden. The user does not have the permission to execute this
        request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: >-
        May contain hints how to solve the error or indicate what went wrong at
        the server. Must not outline any information about the internal
        structure of the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ClientError:
      description: >-
        May contain hints how to solve the error or indicate what was wrong in
        the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: HTTP Conflict 409
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /schemas:
    get:
      tags:
        - Schemas
      summary: 'Get the full list of schemas (ontologies, shapes, vocabularies, JSON schemas, XML schemas).'
      description: 'The schema identifiers, returned by this endpoint can be used to download the schema files. Includes both RDF and non-RDF schema types. Required permission: SCHEMA_READ or ADMIN_ALL'
      operationId: getSchemas
      security:
        - jwt: []
      responses:
        '200':
          description: 'References to ontologies, shapes and vocabularies.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OntologySchema'
        '400':
          $ref: '#/components/responses/ClientError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'          
    post:
      tags:
        - Schemas
      summary: 'Add a new Schema to the catalogue.'
      description: 'This endpoint accepts an ontology (OWL), shape (SHACL), controlled vocabulary (SKOS), JSON Schema, or XML Schema (XSD), which is added to the schemas of the catalogue. Required permission: SCHEMA_CREATE or ADMIN_ALL'
      operationId: addSchema
      security:
        - jwt: []
      requestBody:
        description: 'The file of the new schema: an ontology (OWL), shape (SHACL), controlled vocabulary (SKOS), JSON Schema, or XML Schema (XSD).'
        content:
          application/rdf+xml:
            schema:
              type: string
              description: 'OWL file describing the ontology of the schema or SKOS file describing the controlled vocabulary.'
          application/ld+json:
            schema:
              type: string
              description: 'OWL or SHACL file describing the ontology.'
          application/json:
              schema:
                type: string
                description: 'OWL or SHACL file describing the ontology.'
          text/turtle:
            schema:
              type: string
              description: 'SHACL file describing the ontology of the schema.'
          application/schema+json:
            schema:
              type: string
              description: 'JSON Schema file (Draft 2020-12).'
          application/xml:
            schema:
              type: string
              description: 'XML Schema (XSD) file.'
        required: true
      responses:
        '201':
          description: 'Created'
          headers:
            location:
              description: 'URL of the added Schema.'
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaResult'
        '400':
          $ref: '#/components/responses/ClientError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: 'Unprocessable Entity. The schema content is syntactically invalid.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'
  /schemas/latest:
    get:
      tags:
        - Schemas
      summary: 'Get the latest schema for a given type.'
      description: >-
        Behavior depends on schema type:

        **RDF types** (ontology, shape, vocabulary): Returns the **composite schema** — a merged union
        of all stored schemas of that type. This is meaningful because RDF schemas are additive
        (shapes and ontologies can be combined into a single graph).

        **Non-RDF types** (json, xml): Returns the **most recently uploaded** schema of that type.
        Compositing does not apply because JSON and XML schemas are self-contained and cannot be
        meaningfully merged.
      operationId: getLatestSchema
      parameters:
        - name: type
          in: query
          description: 'Type of the schema.'
          required: true
          schema:
            type: string
            enum:
              - ontology
              - shape
              - vocabulary
              - json
              - xml
        - name: term
          in: query
          description: 'The URI of the term of the requested schema e.g. `https://w3id.org/gaia-x/2511#ServiceOffering`'
          required: false
          schema:
            type: string
      security:
        - jwt: [ ]
      responses:
        '200':
          description: 'The latest schemas for the given type or term.'
          content:
            application/rdf+xml:
              schema:
                type: string
                description: 'OWL file describing the ontology of the schema or SKOS file describing the controlled vocabulary.'
            text/turtle:
              schema:
                type: string
                description: 'SHACL file describing the ontology of the schema.'
            application/schema+json:
              schema:
                type: string
                description: 'JSON Schema file.'
            application/xml:
              schema:
                type: string
                description: 'XML Schema (XSD) file.'
        '400':
          $ref: '#/components/responses/ClientError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /schemas/{schemaId}:
    get:
      tags:
        - Schemas
      summary: Get a specific schema.
      description: 'Required permission: SCHEMA_READ or ADMIN_ALL'
      operationId: getSchema
      security:
        - jwt: []
      parameters:
        - name: schemaId
          in: path
          description: 'Identifier of the Schema.'
          required: true
          schema:
            type: string
        - name: version
          in: query
          description: 'Optional version number. If provided, returns the schema at that specific version.'
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: 'The schema for the given identifier. Depending on the type of the schema, either an ontology, shape graph or controlled vocabulary is returned.'
          content:
            application/rdf+xml:
              schema:
                type: string
                description: 'OWL file describing the ontology of the schema or SKOS file describing the controlled vocabulary.'
            text/turtle:
              schema:
                type: string
                description: 'SHACL file describing the ontology of the schema.'
            application/schema+json:
              schema:
                type: string
                description: 'JSON Schema file.'
            application/xml:
              schema:
                type: string
                description: 'XML Schema (XSD) file.'
        '400':
          $ref: '#/components/responses/ClientError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - Schemas
      summary: 'Replace a schema. This is only allowed for ontologies with the same IRI.'
      description: 'Required permission: SCHEMA_UPDATE or ADMIN_ALL'
      operationId: updateSchema
      parameters:
        - name: schemaId
          in: path
          description: 'Identifier of the Schema.'
          required: true
          schema:
            type: string
      security:
        - jwt: [ ]
      requestBody:
        description: 'The new ontology (OWL file).'
        content:
          application/rdf+xml:
            schema:
              type: string
              description: 'OWL file describing the ontology.'
          application/ld+json:
            schema:
              type: string
              description: 'OWL or SHACL file describing the ontology.'
          application/json:
            schema:
              type: string
              description: 'OWL or SHACL file describing the ontology.'
          text/turtle:
            schema:
              type: string
              description: 'SHACL file describing the ontology of the schema.'
        required: true
      responses:
        '200':
          description: 'Updated.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaResult'
        '400':
          $ref: '#/components/responses/ClientError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - Schemas
      summary: 'Delete a Schema'
      description: 'Required permission: SCHEMA_DELETE or ADMIN_ALL'
      operationId: deleteSchema
      security:
        - jwt: [ ]
      parameters:
        - name: schemaId
          in: path
          description: 'Identifier of the Schema'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Deleted Schema'
        '400':
          $ref: '#/components/responses/ClientError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /schemas/{schemaId}/versions:
    get:
      tags:
        - Schemas
      summary: 'Get version history for a schema.'
      description: 'Returns an ordered list of all versions of the schema. Required permission: SCHEMA_READ or ADMIN_ALL'
      operationId: getSchemaVersions
      security:
        - jwt: []
      parameters:
        - name: schemaId
          in: path
          description: 'Identifier of the Schema.'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'The version history of the schema.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaVersionList'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...